+++ src/Symfony/Bundle/FrameworkBundle/DependencyInjection/Configuration.php
@@ -809,6 +809,7 @@ class Configuration implements ConfigurationInterface
                                     ->scalarNode('provider')
                                         ->info('The service name to use as provider when the specified adapter needs one.')
                                     ->end()
+                                    ->scalarNode('namespace')->end()
                                     ->scalarNode('clearer')->end()
                                 ->end()
                             ->end()